Launching updated, configured IoManager and listener, changed controllers and actions, added comments

+++ FlightGearApi/Domain/Records/GenericConnectionInfo.cs
@@ -2,4 +2,4 @@
 
 namespace FlightGearApi.Domain.Records;
 
-public record GenericConnectionInfo(IoType IoType, int Port, int RefreshesPerSecond, string ProtocolFileName, string Address="127.0.0.1");
+public record GenericConnectionInfo(IoType IoType, int Port, double RefreshesPerSecond, string ProtocolFileName, string Address="127.0.0.1");
